requirements.txt
clip==0.2.0
loguru==0.4.1
matplotlib==3.8.2
numpy==1.26.4
Pillow==10.3.0
seaborn==0.13.2
# torch==2.2.2+cu118
# torchvision==0.17.2+cu118
torch-tb-profiler==0.4.3
tensorboard==2.16.2